chore: In gnokey, show storage deposit fee (#4630)
The total deduction from the account for a transaction is Gas Fee +
Storage Deposit. The gnokey CLI with `-simulate only` shows the
estimated Gas Fee, but not the Storage Deposit. This information is
embedded in the EVENTS, but it is not clear. gnokey should clearly show
the Storage Deposit (which can be much more than the Gas Fee).

* In the gnovm `std` package, add `StorageDepositEvent` and
`StorageUnlockEvent`
* Change `VMKeeper.processStorageDeposit` to report the storage with
`StorageDepositEvent` or `StorageUnlockEvent`, not `GnoEvent`.
* In the tm2 keys `client` `BaseOptions`, add `OnTxSuccess`
* In the tm2 keys `client` package `ExecSignAndBroadcast` and
`execBroadcast`, call `OnTxSuccess` if not nil. (If nil then revert to
the original behavior to print the gas used, etc.)
* In gno.land `keyscli`, add helper function `GetStorageInfo` to search
the events for the storage deposit. Add `PrintTxInfo` to display
transaction result info including storage deposit and total fee.
* In gno.land `keyscli`, when calling `ExecSignAndBroadcast`, set
`OnTxSuccess` to call `PrintTxInfo` like this:
```
STORAGE DELTA:  9 bytes
STORAGE FEE:    900ugnot
TOTAL TX COST:  10900ugnot
```
In case of storage unlock deposit:
```
STORAGE DELTA:  -23 bytes
STORAGE REFUND: 2300ugnot
TOTAL TX COST:  7700ugnot
```

BREAKING CHANGE: `processStorageDeposit` is changed from emitting a
`GnoEvent` to emitting a `StorageDepositEvent` or `StorageUnlockEvent`
with specific fields like `FeeDelta` or `FeeRefund`.

NOTE: We don't add a margin to the displayed storage delta and fee
because these are exact values returned from the node, and also we don't
need to specify `-max-deposit` (unlike `-gas-fee` which should be
specified).

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>